Allow providing a custom bearer token and/or http client implementation for the AppStoreServerAPIClient
