+++ src/main/java/com/apple/itunes/storekit/client/BearerTokenAuthenticatorInterface.java
@@ -1,0 +1,8 @@
+// Copyright (c) 2024 Apple Inc. Licensed under MIT License.
+
+package com.apple.itunes.storekit.client;
+
+@FunctionalInterface
+public interface BearerTokenAuthenticatorInterface {
+    String generateToken();
+}
